requirements.txt
flask
requests
pyrebase4